6 data explore from requests

+++ app/tests/conftest.py
@@ -4,9 +4,11 @@ import pytest
 from sqlalchemy import text
 from app.server import create_app, db
 
+flag_delete_rows = False
+
 
 @pytest.fixture(scope='module')
-def client():
+def client(delete_database_rows=flag_delete_rows):
     app = create_app('test')
     testing_client = app.test_client()
 
@@ -14,8 +16,9 @@ def client():
         db.create_all()
         
     yield testing_client
-        
-    with app.app_context():
-        session = db.session
-        session.execute(text('DELETE FROM data'))
-        session.commit()
+    
+    if delete_database_rows:
+        with app.app_context():
+            session = db.session
+            session.execute(text('DELETE FROM data'))
+            session.commit()
